refactor(analysis)!: Phase 6 S2 — recon registry + config + protocol -> analysis/, sever the contracts->registry knot
The reconciliation registry and its column-mapping config were split across
data/schemas.py and contracts/config.py only because engine/ couldn't host them
(arch_check check 6) and contracts/ validated component names against the
registry. With the runner now in analysis/ (S1), that rationale is gone — and
contracts/ cannot import analysis/, so the registry could not stay below while the
config moved up. Resolution: move the whole reconciliation vocabulary into the
analysis layer.

New src/rwa_calc/analysis/recon_registry.py homes:
- ReconcilableComponent + RECONCILABLE_COMPONENTS(+_BY_NAME) (from data/schemas.py)
- ComponentMapping + LegacyColumnMapping (from contracts/config.py); the unknown-
  component validation now reads the in-module registry (no cross-layer import).
ReconciliationRunnerProtocol moves from contracts/protocols.py into
analysis/reconciliation.py, beside the ReconciliationRunner it types.

Stays put: the RECON_* collapse output-domain tuples in data/schemas.py (read by
engine/aggregator/_collapse.py — engine may not import analysis). Repointed the
consumers: api/__init__ re-exports + api/reconciliation + 5 test files; left
pointer comments at the two old homes. data/schemas.py and contracts/config.py no
longer carry reconciliation config; contracts/ no longer imports the registry.

Gate: ruff + ty clean; arch_check EXIT 0 (no contracts->analysis edge); full suite
7506 passed / 2 skipped. Behaviour-preserving.
