+++ src/rwa_calc/contracts/config.py
@@ -929,91 +929,5 @@ class CalculationConfig:
 # =============================================================================
 RunConfig = CalculationConfig
 
-
-# =============================================================================
-# Parallel-run reconciliation configuration
-# =============================================================================
-# Maps an external (legacy) calculator's output columns onto our canonical
-# reconcilable components (see data/schemas.RECONCILABLE_COMPONENTS) and declares
-# the composite/custom join key. These are configuration, not regulatory data, so
-# they live here alongside the other frozen ``*Config`` dataclasses rather than in
-# data/tables. Tolerance defaults come from the component registry; per-component
-# overrides live on ``ComponentMapping`` (single source).
-
-_RECON_UNITS = ("raw", "decimal", "percent")
-_RECON_TOL_KINDS = ("rel", "abs")
-
-
-@dataclass(frozen=True)
-class ComponentMapping:
-    """How one legacy column maps onto one of our canonical components.
-
-    Attributes:
-        legacy_column: Column name in the legacy output file (pre-normalisation;
-            the loader lowercases + underscores it before lookup).
-        scale: Multiplier applied to the legacy value to reach our units — e.g.
-            legacy RWA in millions uses ``scale=1_000_000``. Amount components only.
-        unit: ``"raw"`` (use as-is), ``"decimal"`` (already 0.20), or ``"percent"``
-            (20.0 → divided by 100). Use for ratio components (pd/lgd/ccf/rw/sf).
-        value_map: Optional legacy→canonical label synonyms for categorical
-            components, e.g. ``{"CORP": "corporate"}``. Keys are matched
-            case-insensitively after normalisation.
-        tol_kind: Optional override of the registry tolerance kind ("rel"|"abs").
-        tol: Optional override of the registry tolerance magnitude.
-    """
-
-    legacy_column: str
-    scale: float = 1.0
-    unit: Literal["raw", "decimal", "percent"] = "raw"
-    value_map: dict[str, str] = field(default_factory=dict)
-    tol_kind: Literal["rel", "abs"] | None = None
-    tol: float | None = None
-
-    def __post_init__(self) -> None:
-        if self.unit not in _RECON_UNITS:
-            raise ValueError(f"unit must be one of {_RECON_UNITS}, got {self.unit!r}")
-        if self.tol_kind is not None and self.tol_kind not in _RECON_TOL_KINDS:
-            raise ValueError(f"tol_kind must be one of {_RECON_TOL_KINDS}, got {self.tol_kind!r}")
-        if self.tol is not None and self.tol < 0:
-            raise ValueError(f"tol must be non-negative, got {self.tol}")
-
-
-@dataclass(frozen=True)
-class LegacyColumnMapping:
-    """Declares how to join and compare a legacy output against our results.
-
-    Attributes:
-        legacy_keys: Ordered key columns in the legacy file forming the join key.
-        our_keys: Ordered key columns on our results frame, positionally aligned
-            with ``legacy_keys``. Defaults to a single ``exposure_reference`` key.
-        components: Canonical-component-name → ``ComponentMapping``. At least one
-            required; names must exist in ``RECONCILABLE_COMPONENTS``.
-    """
-
-    legacy_keys: tuple[str, ...]
-    our_keys: tuple[str, ...] = ("exposure_reference",)
-    components: dict[str, ComponentMapping] = field(default_factory=dict)
-
-    def __post_init__(self) -> None:
-        # Coerce list inputs (e.g. from a TOML loader) to tuples for hashability.
-        object.__setattr__(self, "legacy_keys", tuple(self.legacy_keys))
-        object.__setattr__(self, "our_keys", tuple(self.our_keys))
-
-        if not self.legacy_keys:
-            raise ValueError("legacy_keys must not be empty")
-        if len(self.legacy_keys) != len(self.our_keys):
-            raise ValueError(
-                "legacy_keys and our_keys must be the same length "
-                f"({len(self.legacy_keys)} vs {len(self.our_keys)})"
-            )
-        if not self.components:
-            raise ValueError("at least one component mapping is required")
-
-        from rwa_calc.data.schemas import RECONCILABLE_COMPONENTS_BY_NAME
-
-        unknown = set(self.components) - set(RECONCILABLE_COMPONENTS_BY_NAME)
-        if unknown:
-            valid = sorted(RECONCILABLE_COMPONENTS_BY_NAME)
-            raise ValueError(
-                f"unknown reconciliation components: {sorted(unknown)} (valid: {valid})"
-            )
+# Parallel-run reconciliation configuration (ComponentMapping / LegacyColumnMapping)
+# moved to rwa_calc.analysis.recon_registry (migration Phase 6 - analysis layer).
